pixelart/requirements.txt
flask==3.1.2
flask-cors==6.0.1
google-genai==1.38.0
Pillow==11.3.0
requests==2.32.5
pydantic==2.11.9
vercel==0.3.2